backend/requirements.txt
flask
flask-cors
openai
vectordb
python-dotenv
